requirements.txt
Django==4.2.0
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
joblib==1.2.0
